+++ src/components/content-section/search-failed/SearchFailed.tsx
@@ -3,7 +3,7 @@ import style from './searchFailed.module.scss';
 export const SearchFailed = () => {
   return (
     <div className={style.wrapper}>
-      <p className={style.text}>No Pokemon's with this Search request...</p>
+      <p className={style.text}>No Pokes with this Search request...</p>
     </div>
   );
 };
