+++ src/components/dialog/Dialog.tsx
@@ -1,24 +1,9 @@
-import React, { useState, useCallback } from 'react';
+import React, { useState, useEffect } from 'react';
 import style from './dialog.module.scss';
-
-export enum DialogType {
-  INFO = 'dialogInfo',
-  ERROR = 'dialogError',
-}
+import { DialogState, DialogType } from '../../interface/interface';
+import { bindSetDialog } from './dialogStore';
 
 type DialogProps = Record<string, never>;
-
-type DialogState = {
-  dialogState: boolean;
-  dialogContent: JSX.Element;
-  dialogType: DialogType;
-};
-
-export let openDialog: (
-  dialogContentData: JSX.Element,
-  dialogType: DialogType
-) => void = () => {};
-export let closeDialog: () => void = () => {};
 
 export const Dialog: React.FC<DialogProps> = () => {
   const [dialogState, setDialogState] = useState<DialogState>({
@@ -27,27 +12,17 @@ export const Dialog: React.FC<DialogProps> = () => {
     dialogType: DialogType.INFO,
   });
 
-  const openDialogHandler = useCallback(
-    (dialogContentData: JSX.Element, dialogType: DialogType) => {
-      setDialogState({
-        dialogState: true,
-        dialogContent: dialogContentData,
-        dialogType,
-      });
-    },
-    []
-  );
-
-  const closeDialogHandler = useCallback(() => {
+  function handleClick() {
     setDialogState({
       dialogState: false,
       dialogContent: <p></p>,
       dialogType: DialogType.INFO,
     });
-  }, []);
+  }
 
-  openDialog = openDialogHandler;
-  closeDialog = closeDialogHandler;
+  useEffect(() => {
+    bindSetDialog(setDialogState);
+  }, []);
 
   return (
     <dialog
@@ -56,7 +31,7 @@ export const Dialog: React.FC<DialogProps> = () => {
     >
       <div className={style.dialogContent}>{dialogState.dialogContent}</div>
       <form className={style.dialogButtonWrapper} method="dialog">
-        <button className={style.dialogButton} onClick={closeDialogHandler}>
+        <button className={style.dialogButton} onClick={handleClick}>
           Close
         </button>
       </form>
