+++ src/components/not-found-page/NotFoundPage.tsx
@@ -1,7 +1,8 @@
 import { useNavigate } from 'react-router-dom';
-import { Button, ButtonType } from '../button/Button';
+import { Button } from '../button/Button';
 import style from './newFoundPage.module.scss';
 import ErrPoke from '/404page.png';
+import { ButtonType } from '../../interface/interface';
 
 export const NotFoundPage = () => {
   const navigate = useNavigate();
